+++ libs/components/text-editor/src/lib/modules/text-editor/text-editor.component.ts
@@ -26,6 +26,7 @@ import { SkyFormsUtility } from '../shared/forms-utility';
 
 import { FONT_LIST_DEFAULTS } from './defaults/font-list-defaults';
 import { FONT_SIZE_LIST_DEFAULTS } from './defaults/font-size-list-defaults';
+import { LINK_WINDOW_OPTIONS_DEFAULTS } from './defaults/link-window-options-defaults';
 import { MENU_DEFAULTS } from './defaults/menu-defaults';
 import { STYLE_STATE_DEFAULTS } from './defaults/style-state-defaults';
 import { TOOLBAR_ACTION_DEFAULTS } from './defaults/toolbar-action-defaults';
@@ -36,6 +37,7 @@ import { SkyTextEditorService } from './services/text-editor.service';
 import { SkyTextSanitizationService } from './services/text-sanitization.service';
 import { SkyTextEditorToolbarComponent } from './toolbar/text-editor-toolbar.component';
 import { SkyTextEditorFont } from './types/font-state';
+import { SkyTextEditorLinkWindowOptionsType } from './types/link-window-options-type';
 import { SkyTextEditorMenuType } from './types/menu-type';
 import { SkyTextEditorStyleState } from './types/style-state';
 import { SkyTextEditorMergeField } from './types/text-editor-merge-field';
@@ -217,7 +219,7 @@ export class SkyTextEditorComponent
   }
 
   /**
-   * The actions to include in the toolbar and determines their order.
+   * The actions to include in the toolbar in the specified order.
    * @default [ 'font-family', 'font-size', 'font-style', 'color', 'list', 'link ]
    */
   @Input()
@@ -229,6 +231,21 @@ export class SkyTextEditorComponent
 
   public get toolbarActions(): SkyTextEditorToolbarActionType[] {
     return this.#_toolbarActions;
+  }
+
+  /**
+   * The target window options for adding a hyperlink.
+   * @default [ 'new', 'existing' ]
+   */
+  @Input()
+  public set linkWindowOptions(
+    value: SkyTextEditorLinkWindowOptionsType[] | undefined
+  ) {
+    this.#_linkWindowOptions = value ?? LINK_WINDOW_OPTIONS_DEFAULTS;
+  }
+
+  public get linkWindowOptions(): SkyTextEditorLinkWindowOptionsType[] {
+    return this.#_linkWindowOptions;
   }
 
   @ViewChild('iframe')
@@ -311,6 +328,7 @@ export class SkyTextEditorComponent
   #_mergeFields: SkyTextEditorMergeField[] = [];
   #_menus = MENU_DEFAULTS;
   #_toolbarActions: SkyTextEditorToolbarActionType[] = TOOLBAR_ACTION_DEFAULTS;
+  #_linkWindowOptions = LINK_WINDOW_OPTIONS_DEFAULTS;
   #_disabled = false;
   #_initialStyleState = Object.assign({}, STYLE_STATE_DEFAULTS);
   #_placeholder = '';