+++ libs/components/text-editor/src/lib/modules/text-editor/url-modal/text-editor-url-modal-context.ts
@@ -1,3 +1,5 @@
+import { SkyTextEditorLinkWindowOptionsType } from '../types/link-window-options-type';
+
 import { UrlModalResult } from './text-editor-url-modal-result';
 
 /**
@@ -5,4 +7,5 @@ import { UrlModalResult } from './text-editor-url-modal-result';
  */
 export class SkyUrlModalContext {
   public urlResult: UrlModalResult | undefined;
+  public linkWindowOptions: SkyTextEditorLinkWindowOptionsType[] | undefined;
 }